Finalized much of photo upload process
Cleaned up imports.
Wrapped up a lot of backend and fronted.
Boundaries still not coming with tract objects.

+++ vhl/src/main/java/info/cinow/dto/PhotoDto.java
@@ -14,5 +14,7 @@ public class PhotoDto {
 
     private String description;
 
-    private Long tractId;
+    private Integer tractId;
+
+    // TODO: link to image on S3 bucket
 }
